backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
SQLAlchemy==2.0.21
alembic==1.12.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
httpx==0.24.1
python-jose==3.3.0

